chart: use <Tooltip /> component
