+++ client/components/chart/style.scss
@@ -354,94 +354,7 @@
 // Chart tooltip
 
 .chart__tooltip {
-	z-index: z-index( 'root', '.chart__tooltip' );
-	position: absolute;
-	/* default offset for edge-cases: https://github.com/component/tip/pull/12 */
-	top: 0;
-	left: 0 #{"/*rtl:ignore*/"};
-	right: auto #{"/*rtl:ignore*/"};
-
-	&.is-bottom-left {
-		.popover__arrow {
-			border-color: $transparent;
-			position: absolute;
-
-			&::before {
-				content: '';
-				border: 4px solid $transparent;
-				border-bottom-color: darken( $gray, 30% );
-				position: absolute;
-					left: 231px  #{"/*rtl:ignore*/"};
-					top: -3px;
-
-				.rtl & {
-					left: -18px  #{"/*rtl:ignore*/"};
-					top: -3px;
-				}
-			}
-		}
-	}
-
-	&.is-bottom-right,
-	&.is-bottom {
-		.popover__arrow {
-			border-color: $transparent;
-			position: absolute;
-
-			&::before {
-				content: '';
-				border: 4px solid $transparent;
-				border-bottom-color: darken( $gray, 30% );
-				position: absolute;
-					left: 11px  #{"/*rtl:ignore*/"};
-					top: -3px;
-
-				.rtl & {
-					left: -238px  #{"/*rtl:ignore*/"};
-					top: -3px;
-				}
-			}
-		}
-	}
-
-	&.is-top {
-		.popover__arrow {
-			border-color: $transparent;
-			position: absolute;
-
-			&::before {
-				content: '';
-				border: 4px solid $transparent;
-				border-top-color: darken( $gray, 30% );
-				position: absolute;
-					left: 118px #{"/*rtl:ignore*"};
-					bottom: -40px;
-			}
-		}
-	}
-
-	&.is-bottom {
-		.popover__arrow {
-			&::before {
-				left: 86px  #{"/*rtl:ignore*/"};
-				top: -4px;
-
-				.rtl & {
-					left: -93px  #{"/*rtl:ignore*/"};
-				}
-			}
-		}
-	}
-
 	.popover__inner {
-		border: 0;
-		box-shadow: none;
-		border-radius: 2px;
-		color: $white;
-		background: darken( $gray, 30% );
-		font-size: 11px;
-		padding: 6px 10px;
-		margin-top: 4px;
 		width: 230px;
 		text-align: left;
